requirements.txt
apache-airflow==3.0.6
scikit-learn
pandas
numpy